pets displayed on dashboard

+++ server/public/scripts/controllers/owners.controller.js
@@ -1,43 +1,47 @@
-app.controller('OwnersController', ['$http', function($http) {
+app.controller('OwnersController', ['$http', function ($http) {
     let vm = this;
     vm.owners = [];
 
-    vm.getOwners = function() {
+    vm.getOwners = function () {
         $http({
             method: 'GET',
             url: '/owner'
-        }).then( function(response) {
+        }).then(function (response) {
             vm.owners = response.data;
-        }).catch( function(error) {
+        }).catch(function (error) {
             console.log('error:', error);
             alert('failed to retrieve owner data');
         });
     }//end getOwners
 
-    vm.addOwner = function(name) {
+    vm.addOwner = function (name) {
         $http({
             method: 'POST',
             url: '/owner',
-            data: {name: name}
-        }).then( function(response) {
+            data: { name: name }
+        }).then(function (response) {
             vm.getOwners();
-        }).catch( function(error) {
+        }).catch(function (error) {
             console.log('error:', error);
             alert('failed to reach database');
         });
     }//end addOwner
 
-    vm.removeOwner = function(owner) {
-        $http({
-            method: 'DELETE',
-            url: '/owner',
-            params: {id: owner.id}
-        }).then( function(response) {
-            vm.getOwners();
-        }).catch( function(error) {
-            console.log('error:', error);
-            alert('failed to remove owner');
-        });
+    vm.removeOwner = function (owner) {
+        if (owner.count > 0) {
+            alert('you cannot delete an owner with registered pets');
+        } else {
+            $http({
+                method: 'DELETE',
+                url: '/owner',
+                params: { id: owner.id }
+            }).then(function (response) {
+                vm.getOwners();
+            }).catch(function (error) {
+                console.log('error:', error);
+                alert('failed to remove owner');
+            });
+        }
     }//end removeOwner
 
     vm.getOwners();
